HUB-737: Removed redirect to idp warning page
Renamed and refactored redirect to idp warning controller to
choose_a_certified_company_redirect_controller which contains
logic to redirect from the Certified company choice page for both
(LOA1) and (LOA2) straight to idp registration when idp is viewable
and is available for registration. Removed associated specs for redirect_to_idp_warning
controller and views also removed questions pages and associated test
