--- app/views/redirect_to_idp_question/idp_wont_work_for_you.html.erb
@@ -1,15 +1,0 @@
-<%= page_title 'hub.idp_wont_work_for_you_one_doc.title', {:idp_name => @idp.display_name} %>
-<% content_for :feedback_source, 'IDP_WILL_NOT_WORK_FOR_YOU' %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l"><%= t 'hub.idp_wont_work_for_you_one_doc.heading', idp_name: @idp.display_name %></h1>
-        <%= raw @idp.interstitial_explanation %>
-    <div>
-      <p>
-        <%= t 'hub.idp_wont_work_for_you_one_doc.choose_another_company' %>
-      </p>
-      <%= button_link_to t('hub.idp_wont_work_for_you_one_doc.choose_another_certified_company_link'), choose_a_certified_company_path, id: 'choose-another-certified-company-link', class: 'govuk-button' %>
-    </div>
-  </div>
-</div>
