+++ spec/controllers/choose_a_certified_company_loa1_controller_spec.rb
@@ -50,8 +50,11 @@ describe ChooseACertifiedCompanyLoa1Controller do
 
   context "#select_idp" do
     before :each do
+      stub_api_select_idp
       set_session_and_cookies_with_loa("LEVEL_1")
+      stub_api_idp_list_for_sign_in [stub_idp_loa1, stub_idp_loa1_with_interstitial]
       stub_api_idp_list_for_registration([stub_idp_loa1, stub_idp_loa1_with_interstitial], "LEVEL_1")
+      allow_any_instance_of(UserCookiesPartialController).to receive(:ab_test_with_alternative_name).and_return(nil)
     end
 
     it "resets interstitial answer to no value when IDP is selected" do
@@ -77,13 +80,7 @@ describe ChooseACertifiedCompanyLoa1Controller do
       stub_api_idp_list_for_registration([stub_idp_no_interstitial], "LEVEL_1")
       post :select_idp, params: { locale: "en", entity_id: "http://idcorp-two.com" }
 
-      expect(subject).to redirect_to redirect_to_idp_warning_path
-    end
-
-    it "redirects to IDP question page for LOA1 users when IDP flag is enabled" do
-      post :select_idp, params: { locale: "en", entity_id: "http://idcorp-loa1-with-interstitial.com" }
-
-      expect(subject).to redirect_to redirect_to_idp_question_path
+      expect(subject).to redirect_to redirect_to_idp_register_path
     end
 
     it "returns 404 page if IDP is non-existent" do
