+++ spec/support/list_size_matcher.rb
@@ -1,0 +1,3 @@
+RSpec::Matchers.define :a_list_of_size do |x|
+  match { |actual| actual.length == x }
+end
