+++ stub/idp-rules/stub-idp-demo.yml
@@ -5,7 +5,7 @@ segments:
     recommended: []
     unlikely: []
   non_protected:
-    recommended: ["other"]
+    recommended: []
     unlikely: []
 
 capabilities:
