+++ stub/idp-rules/stub-idp-loa1-with-interstital.yml
@@ -6,7 +6,7 @@ segments:
     recommended: []
     unlikely: []
   non_protected:
-    recommended: ["other"]
+    recommended: []
     unlikely: []
 
 capabilities:
