+++ stub/idp-rules/stub-idp-loa1.yml
@@ -5,7 +5,7 @@ segments:
     recommended: []
     unlikely: []
   non_protected:
-    recommended: ["other"]
+    recommended: []
     unlikely: []
 
 capabilities:
